requirements.txt
# Web Server
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
aiofiles
google-search-results
streamlit
request 
w
# AI Framework & Google Integration
langchain
langchain-community
langchain-google-genai
google-ai-generativelanguage

# Document Processing & Vector Store
pypdf
pdfplumber
chromadb